+++ Aelia's Apprentice/Assets/Scripts/BreakableObject.cs
@@ -4,6 +4,6 @@ public class BreakableObject : MonoBehaviour, IDamageable
 {
     public void TakeDamage()
     {
-        Debug.Log(gameObject + ": I have been damaged");
+        Debug.Log(gameObject + ": I am breaking");
     }
 }